requirements.txt
pillow
numpy
cython
pathlib
scipy
setuptools
streamlit
streamlit-echarts
matplot
